remove vmaas listener from listener
