+++ Dockerfile-listener.rhel7
@@ -20,10 +20,10 @@ RUN chown -R insights $APPBASEDIR
 
 RUN $APPBASEDIR/scl-enable.sh pip install --upgrade pip
 
-RUN $APPBASEDIR/scl-enable.sh pip install psycopg2-binary tornado requests insights-core prometheus_client "aiokafka==0.5.0"
+RUN $APPBASEDIR/scl-enable.sh pip install psycopg2-binary requests insights-core prometheus_client "aiokafka==0.5.0"
 
 USER insights
 
-EXPOSE 8000 8086
+EXPOSE 8086
 
 CMD ["sh", "-c", "$APPBASEDIR/scl-enable.sh $APPBASEDIR/entrypoint.sh"]
