+++ conf/listener.env
@@ -1,4 +1,3 @@
-VMAAS_WEBSOCKET_HOST=vmaas_websocket:8082
 LOGGING_TYPE=DEVEL
 LOGGING_LEVEL=INFO
 MESSAGE_SERVICE=ve_kafka:9092
