+++ listener/entrypoint.sh
@@ -1,4 +1,4 @@
 #!/bin/sh
 
 cd $(dirname $0)
-exec ./wait-for-services.sh python3 -m listener.listener
+exec ./wait-for-services.sh python3 -m listener.upload_listener
