--- listener/listener.py
@@ -1,26 +1,0 @@
-#!/usr/bin/env python3
-"""Main module for listener."""
-
-from multiprocessing import Process
-import os
-from prometheus_client import start_http_server
-
-from .vmaas_listener import main as vmaas_listener
-from .upload_listener import main as upload_listener
-
-prometheus_port = os.getenv('PROMETHEUS_PORT', '8086') # pylint: disable=invalid-name
-
-def main():
-    """Main entrypoint for listener."""
-    start_http_server(int(prometheus_port))
-
-    vmaas_listener_process = Process(target=vmaas_listener)
-    vmaas_listener_process.start()
-    upload_listener_process = Process(target=upload_listener)
-    upload_listener_process.start()
-
-    vmaas_listener_process.join()
-    upload_listener_process.join()
-
-if __name__ == '__main__':
-    main()
