+++ listener/upload_listener.py
@@ -1,18 +1,20 @@
 """Module implementing kafka listener."""
 
 import json
+import os
 import tempfile
 
-from prometheus_client import Counter
+from prometheus_client import Counter, start_http_server
 import requests
 
 from common import mqueue
 from common.database_handler import DatabaseHandler, init_db
 from common.logging import init_logging, get_logger
 from .archive_parser import ArchiveParser
-from .utils import VERSION, format_vmaas_request
 
 LOGGER = get_logger(__name__)
+
+PROMETHEUS_PORT = os.getenv('PROMETHEUS_PORT', '8087')
 
 NEW_SYSTEM = Counter('ve_listener_upl_new_system', '# of new systems inserted')
 UPDATE_SYSTEM = Counter('ve_listener_upl_update_system', '# of systems updated')
@@ -20,6 +22,16 @@ PROCESS_UPLOAD = Counter('ve_listener_upl_uploads_processed', '# of uploaded arc
 MISSING_ID = Counter('ve_listener_upl_missing_inventory_id', '# of upload-msgs missing inventory_id')
 ARCHIVE_PARSE_FAILURE = Counter('ve_listener_upl_archive_excpetions', '# of exceptions during archive-processing')
 ARCHIVE_NO_RPMDB = Counter('ve_listener_upl_no_rpmdb', '# of systems ignored due to missing rpmdb')
+
+
+def format_vmaas_request(package_list, repo_list=None, modules_list=None):
+    """Wrap package and repo list into the vmaas request format."""
+    vmaas_request = {"package_list": package_list}
+    if repo_list:
+        vmaas_request["repository_list"] = repo_list
+    if modules_list:
+        vmaas_request["modules_list"] = modules_list
+    return json.dumps(vmaas_request)
 
 
 def db_import_system(conn, inventory_id, rh_account, s3_url, vmaas_json, satellite_managed):
@@ -46,9 +58,10 @@ def db_import_system(conn, inventory_id, rh_account, s3_url, vmaas_json, satelli
 
 def main():
     """Main kafka listener entrypoint."""
+    start_http_server(int(PROMETHEUS_PORT))
     init_logging()
     init_db()
-    LOGGER.info("Starting upload listener (version %s).", VERSION)
+    LOGGER.info("Starting upload listener.")
     # get DB connection
     conn = DatabaseHandler.get_connection()
     upload_queue = mqueue.MQReader(mqueue.UPLOAD_TOPIC)
