--- listener/utils.py
@@ -1,22 +1,0 @@
-"""Utility functions for listener."""
-
-import os
-import json
-from uuid import uuid4
-
-VERSION = os.getenv("VERSION", "master")
-
-
-def uuid():
-    """Return random uuid."""
-    return str(uuid4())
-
-
-def format_vmaas_request(package_list, repo_list=None, modules_list=None):
-    """Wrap package and repo list into the vmaas request format."""
-    vmaas_request = {"package_list": package_list}
-    if repo_list:
-        vmaas_request["repository_list"] = repo_list
-    if modules_list:
-        vmaas_request["modules_list"] = modules_list
-    return json.dumps(vmaas_request)
